This is a squashed effort to get lita to recconect to persited rooms.
Now when Lita restarts the hipchat adapter will look at the persisted rooms
and attempt to rejoin them.

First pass at re-joining persisted rooms.

This is my first pass at re-joining persisted rooms. I need to figure out
how to test the new persisted rooms functionality.

Using redis to retreive persisted rooms.

I made it so that the adapter reads the persisted rooms out of redis.
Also added specs around the method to make sure it behaves as expected.

Add nil check in case persisted rooms is empty.

I wanted to make sure we gracefully handle the case where persisted_rooms
doesn't exist in redis so I added a nil check and tested for it.

reverting to previous implementation and more specs.

I went back to the old style of implementation but this time flushed it out with
more specs. I made it so the adapter asks the robot for the list of persisted
rooms and then tries to join them.

removing pry require.

I had this in to do debugging but I forgot to pull it out.