Removing rescue from persisted_rooms join.
There was a bug that would allow malformed rooms to get into the persisted rooms
list so I put a check in that made sure those rooms didn't get in there.
Since we fixed that bug in Lita I am removing the unecessary check.

Removing test for bad room name.

This test was no longer necessary since we fixed the bug in lita and removed
the coresponding rescue.